+++ package.json
@@ -33,6 +33,7 @@
     "@vitest/coverage-v8": "^3.2.4",
     "autoprefixer": "10.4.16",
     "gh-pages": "^6.1.1",
+    "ignore-loader": "^0.1.2",
     "jsdom": "^26.1.0",
     "postcss": "8.4.32",
     "tailwindcss": "3.4.0",
